+++ app/javascript/shared/constants/messages.js
@@ -1,10 +1,3 @@
-import {
-  DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE,
-  resolveMaximumFileUploadSize,
-} from 'shared/helpers/FileUploadLimitHelper';
-
-export { DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE, resolveMaximumFileUploadSize };
-
 export const MESSAGE_STATUS = {
   FAILED: 'failed',
   SENT: 'sent',
@@ -40,13 +33,6 @@ export const CONVERSATION_PRIORITY_ORDER = {
   medium: 2,
   low: 1,
 };
-
-const globalWindow = typeof window !== 'undefined' ? window : {};
-
-// Size in mega bytes
-export const MAXIMUM_FILE_UPLOAD_SIZE = resolveMaximumFileUploadSize(
-  globalWindow?.globalConfig?.MAXIMUM_FILE_UPLOAD_SIZE
-);
 
 export const ALLOWED_FILE_TYPES =
   'image/*,' +
