+++ app/javascript/shared/helpers/specs/FileHelper.spec.js
@@ -1,7 +1,9 @@
 import {
+  DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE,
   formatBytes,
   fileSizeInMegaBytes,
   checkFileSizeLimit,
+  resolveMaximumFileUploadSize,
 } from '../FileHelper';
 
 describe('#File Helpers', () => {
@@ -19,6 +21,7 @@ describe('#File Helpers', () => {
       expect(formatBytes(10000000)).toBe('9.54 MB');
     });
   });
+
   describe('fileSizeInMegaBytes', () => {
     it('should return zero if 0 is passed', () => {
       expect(fileSizeInMegaBytes(0)).toBe(0);
@@ -27,12 +30,35 @@ describe('#File Helpers', () => {
       expect(fileSizeInMegaBytes(20000000)).toBeCloseTo(19.07, 2);
     });
   });
+
   describe('checkFileSizeLimit', () => {
     it('should return false if file with size 62208194 and file size limit 40 are passed', () => {
       expect(checkFileSizeLimit({ file: { size: 62208194 } }, 40)).toBe(false);
     });
     it('should return true if file with size 62208194 and file size limit 40 are passed', () => {
       expect(checkFileSizeLimit({ file: { size: 199154 } }, 40)).toBe(true);
+    });
+  });
+
+  describe('resolveMaximumFileUploadSize', () => {
+    it('should return default when value is undefined', () => {
+      expect(resolveMaximumFileUploadSize(undefined)).toBe(
+        DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE
+      );
+    });
+
+    it('should return default when value is not a positive number', () => {
+      expect(resolveMaximumFileUploadSize('not-a-number')).toBe(
+        DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE
+      );
+      expect(resolveMaximumFileUploadSize(-5)).toBe(
+        DEFAULT_MAXIMUM_FILE_UPLOAD_SIZE
+      );
+    });
+
+    it('should parse numeric strings and numbers', () => {
+      expect(resolveMaximumFileUploadSize('50')).toBe(50);
+      expect(resolveMaximumFileUploadSize(75)).toBe(75);
     });
   });
 });
